refactor(agent-task): extract timeout artifact discovery
