+++ src/core/mod.rs
@@ -8,6 +8,7 @@ pub mod agent_task_promotion;
 pub mod agent_task_provider;
 pub mod agent_task_schedule;
 pub mod agent_task_scheduler;
+pub(crate) mod agent_task_timeout_artifacts;
 pub mod api_jobs;
 pub mod artifact_manifest;
 pub(crate) mod artifact_metadata;
